requirements.txt
fastapi==0.95.2
uvicorn==0.20.0
requests==2.28.2
influxdb-client==1.36.1
redis==4.5.1
httpx==0.23.3
pycryptodome==3.17.0
itsdangerous==2.1.2
bcrypt==4.0.1
python-multipart==0.0.5
pydantic==1.10.7
psutil==7.0.0
ujson==5.7.0
httptools==0.5.0